\name{MOWAttrSelection}
\alias{MOWAttrSelection}
\alias{MOWAttrSelection}
\docType{package}
\title{
Attribute selection
}
\description{
Attribute selection with wrappers
}
\details{
\tabular{ll}{
Package: \tab MOWAttrSelection\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-01-24\cr
License: \tab What license is it under?\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
MCZ PK
}
\keyword{ package }
