% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MOWAttrSelection.R
\name{selectAttributes}
\alias{selectAttributes}
\title{Select attributes based on train data and train data labels (results)}
\usage{
selectAttributes(trainData, trainLabels, searchMethod = "hill")
}
\arguments{
\item{trainData}{Train data. Must be a data-frame.}

\item{trainLabels}{Train labels (results). Must be a data-frame.}

\item{searchMethod}{Search method used in attributes selections. Possible choices are 'hill'
for hill climbing search (default option) and 'bestFirst' for best first search}
}
\value{
Vector with selected attributes
}
\description{
Select attributes based on train data and train data labels (results)
}
\examples{
#select attributes using test data, and test data labels (results)
attrVector <- selectAttributes(trainData, trainLabels, searchMethod)
use selected attributes in your train data
#predict(model, testData[attrVector])
}

